# dynamodbattribute: cache struct type info, refactor fieldsByIndex