--- a/service/dynamodb/dynamodbattribute/encode_test.go
+++ b/service/dynamodb/dynamodbattribute/encode_test.go
@@ -276,30 +276,26 @@ func TestEncoderFieldByIndex(t *testing.T) {
 		Outer  struct{ *Middle }
 	)
 
-	t.Run("nil embedded struct", func(t *testing.T) {
-		outer := Outer{}
+	// nil embedded struct
+	outer := Outer{}
+	outerFields := unionStructFields(reflect.TypeOf(outer), MarshalOptions{})
+	innerField, _ := outerFields.FieldByName("Inner")
 
-		outerFields := unionStructFields(reflect.TypeOf(outer), MarshalOptions{})
-		innerField, _ := outerFields.FieldByName("Inner")
+	_, found := encoderFieldByIndex(reflect.ValueOf(&outer).Elem(), innerField.Index)
+	if found != false {
+		t.Error("expected found to be false when embedded struct is nil")
+	}
 
-		_, found := encoderFieldByIndex(reflect.ValueOf(&outer).Elem(), innerField.Index)
-		if found != false {
-			t.Error("expected found to be false when embedded struct is nil")
-		}
-	})
+	// non-nil embedded struct
+	outer = Outer{Middle: &Middle{Inner: 3}}
+	outerFields = unionStructFields(reflect.TypeOf(outer), MarshalOptions{})
+	innerField, _ = outerFields.FieldByName("Inner")
 
-	t.Run("non-nil embedded struct", func(t *testing.T) {
-		outer := Outer{Middle: &Middle{Inner: 3}}
-
-		outerFields := unionStructFields(reflect.TypeOf(outer), MarshalOptions{})
-		innerField, _ := outerFields.FieldByName("Inner")
-
-		f, found := encoderFieldByIndex(reflect.ValueOf(&outer).Elem(), innerField.Index)
-		if found != true {
-			t.Error("expected found to be true")
-		}
-		if f.Kind() != reflect.Int || f.Int() != int64(outer.Inner) {
-			t.Error("expected f to be of kind Int with value equal to outer.Inner")
-		}
-	})
+	f, found := encoderFieldByIndex(reflect.ValueOf(&outer).Elem(), innerField.Index)
+	if found != true {
+		t.Error("expected found to be true")
+	}
+	if f.Kind() != reflect.Int || f.Int() != int64(outer.Inner) {
+		t.Error("expected f to be of kind Int with value equal to outer.Inner")
+	}
 }