--- a/service/dynamodb/dynamodbattribute/marshaler_test.go
+++ b/service/dynamodb/dynamodbattribute/marshaler_test.go
@@ -555,7 +555,7 @@ func compareObjects(t *testing.T, expected interface{}, actual interface{}) {
 }
 
 func BenchmarkMarshal(b *testing.B) {
-	d := simpleMarshalStruct{
+	simple := simpleMarshalStruct{
 		String:  "abc",
 		Int:     123,
 		Uint:    123,
@@ -564,12 +564,85 @@ func BenchmarkMarshal(b *testing.B) {
 		Bool:    true,
 		Null:    nil,
 	}
-	for i := 0; i < b.N; i++ {
-		_, err := Marshal(d)
-		if err != nil {
-			b.Fatal("unexpected error", err)
+	b.Run("one composite member", func(b *testing.B) {
+		type MyCompositeStruct struct {
+			A simpleMarshalStruct `dynamodbav:"a"`
 		}
-	}
+		b.RunParallel(func(pb *testing.PB) {
+			for pb.Next() {
+				if _, err := Marshal(MyCompositeStruct{
+					A: simple,
+				}); err != nil {
+					b.Error("unexpected error:", err)
+				}
+			}
+		})
+	})
+
+	b.Run("two composite members", func(b *testing.B) {
+		type MyCompositeStruct struct {
+			A simpleMarshalStruct `dynamodbav:"a"`
+			B simpleMarshalStruct `dynamodbav:"b"`
+		}
+		b.RunParallel(func(pb *testing.PB) {
+			for pb.Next() {
+				if _, err := Marshal(MyCompositeStruct{
+					A: simple,
+					B: simple,
+				}); err != nil {
+					b.Error("unexpected error:", err)
+				}
+			}
+		})
+	})
+}
+
+func BenchmarkUnmarshal(b *testing.B) {
+	myStructAVMap, _ := Marshal(simpleMarshalStruct{
+		String:  "abc",
+		Int:     123,
+		Uint:    123,
+		Float32: 123.321,
+		Float64: 123.321,
+		Bool:    true,
+		Null:    nil,
+	})
+
+	b.Run("one composite member", func(b *testing.B) {
+		type MyCompositeStructOne struct {
+			A simpleMarshalStruct `dynamodbav:"a"`
+		}
+		var out MyCompositeStructOne
+		avMap := map[string]*dynamodb.AttributeValue{
+			"a": myStructAVMap,
+		}
+		b.RunParallel(func(pb *testing.PB) {
+			for pb.Next() {
+				if err := Unmarshal(&dynamodb.AttributeValue{M: avMap}, &out); err != nil {
+					b.Error("unexpected error:", err)
+				}
+			}
+		})
+	})
+
+	b.Run("two composite members", func(b *testing.B) {
+		type MyCompositeStructTwo struct {
+			A simpleMarshalStruct `dynamodbav:"a"`
+			B simpleMarshalStruct `dynamodbav:"b"`
+		}
+		var out MyCompositeStructTwo
+		avMap := map[string]*dynamodb.AttributeValue{
+			"a": myStructAVMap,
+			"b": myStructAVMap,
+		}
+		b.RunParallel(func(pb *testing.PB) {
+			for pb.Next() {
+				if err := Unmarshal(&dynamodb.AttributeValue{M: avMap}, &out); err != nil {
+					b.Error("unexpected error:", err)
+				}
+			}
+		})
+	})
 }
 
 func Test_Encode_YAML_TagKey(t *testing.T) {