Currently struct type information is recomputed on every call to `dynamodbattribute.Marshal` and `dynamodbattribute.Unmarshal`.  To avoid this repeated work, I've added a `sync.Map`-backed cache to the reflection pass so that struct type info is only computed once.  In addition to caching fields per type, I've also included an index for each type that maps field names to fields to avoid linear searches by name in the Decoder.  On the whole, this approach basically the same as what the stdlib's `encoding/json` [does to solve the same problem](https://github.com/golang/go/blob/25a14b19abd3b9e16f47c6249fda1998431ce5be/src/encoding/json/encode.go#L1379).

`dynamodbattribute.fieldsByIndex` has been refactored into two functions, `decoderFieldsByIndex` and `encoderFieldsByIndex`.  This approach eliminates the need for the `OnEmbeddedNilStruct` callback and as a result no longer requires that the instance of `reflect.Value` in the search loop be moved to the heap.    

I've added a couple of benchmarks to quantify the impact of the changes.  Here's `benchcmp` diff of the results before and after:

```
goos: darwin
goarch: amd64

benchmark                                      old ns/op     new ns/op     delta
BenchmarkMarshal/one_composite_member          8919          3486          -60.91%
BenchmarkMarshal/one_composite_member-2        4939          1972          -60.07%
BenchmarkMarshal/one_composite_member-4        2987          1245          -58.32%
BenchmarkMarshal/one_composite_member-8        2949          1293          -56.15%
BenchmarkMarshal/two_composite_members         16569         6599          -60.17%
BenchmarkMarshal/two_composite_members-2       9201          3693          -59.86%
BenchmarkMarshal/two_composite_members-4       5762          2380          -58.69%
BenchmarkMarshal/two_composite_members-8       5678          2467          -56.55%
BenchmarkUnmarshal/one_composite_member        7219          1833          -74.61%
BenchmarkUnmarshal/one_composite_member-2      3881          995           -74.36%
BenchmarkUnmarshal/one_composite_member-4      2330          571           -75.49%
BenchmarkUnmarshal/one_composite_member-8      2349          572           -75.65%
BenchmarkUnmarshal/two_composite_members       14010         3223          -77.00%
BenchmarkUnmarshal/two_composite_members-2     7598          1707          -77.53%
BenchmarkUnmarshal/two_composite_members-4     4455          1022          -77.06%
BenchmarkUnmarshal/two_composite_members-8     4591          1041          -77.33%

benchmark                                      old allocs     new allocs     delta
BenchmarkMarshal/one_composite_member          70             31             -55.71%
BenchmarkMarshal/one_composite_member-2        70             31             -55.71%
BenchmarkMarshal/one_composite_member-4        70             31             -55.71%
BenchmarkMarshal/one_composite_member-8        70             31             -55.71%
BenchmarkMarshal/two_composite_members         132            57             -56.82%
BenchmarkMarshal/two_composite_members-2       132            57             -56.82%
BenchmarkMarshal/two_composite_members-4       132            57             -56.82%
BenchmarkMarshal/two_composite_members-8       132            57             -56.82%
BenchmarkUnmarshal/one_composite_member        42             3              -92.86%
BenchmarkUnmarshal/one_composite_member-2      42             3              -92.86%
BenchmarkUnmarshal/one_composite_member-4      42             3              -92.86%
BenchmarkUnmarshal/one_composite_member-8      42             3              -92.86%
BenchmarkUnmarshal/two_composite_members       79             4              -94.94%
BenchmarkUnmarshal/two_composite_members-2     79             4              -94.94%
BenchmarkUnmarshal/two_composite_members-4     79             4              -94.94%
BenchmarkUnmarshal/two_composite_members-8     79             4              -94.94%

benchmark                                      old bytes     new bytes     delta
BenchmarkMarshal/one_composite_member          4560          2456          -46.14%
BenchmarkMarshal/one_composite_member-2        4560          2456          -46.14%
BenchmarkMarshal/one_composite_member-4        4559          2456          -46.13%
BenchmarkMarshal/one_composite_member-8        4560          2456          -46.14%
BenchmarkMarshal/two_composite_members         8608          4432          -48.51%
BenchmarkMarshal/two_composite_members-2       8608          4432          -48.51%
BenchmarkMarshal/two_composite_members-4       8607          4432          -48.51%
BenchmarkMarshal/two_composite_members-8       8608          4432          -48.51%
BenchmarkUnmarshal/one_composite_member        2336          240           -89.73%
BenchmarkUnmarshal/one_composite_member-2      2336          240           -89.73%
BenchmarkUnmarshal/one_composite_member-4      2336          240           -89.73%
BenchmarkUnmarshal/one_composite_member-8      2336          240           -89.73%
BenchmarkUnmarshal/two_composite_members       4432          272           -93.86%
BenchmarkUnmarshal/two_composite_members-2     4432          272           -93.86%
BenchmarkUnmarshal/two_composite_members-4     4432          272           -93.86%
BenchmarkUnmarshal/two_composite_members-8     4432          272           -93.86%
```