@jasdel 2020-01-07
Thanks for taking the time to create this PR. The performance improvements are nice to see, and would be good to include in the SDK.

I think the primary issue blocking accepting this PR is that sync.Map was adding in Go 1.9, and the SDK supports all the way back to Go 1.5. If possible it would be great to include this change in the SDK for Go 1.9 and above.

I do have a bit of concern about the potential issue with unbounded memory the sync.Map could contain. This could be seen as a memory leak since there is no way to flush the map when it is no longer needed. I wasn't able to find any discussion about the `encoding/json` update with regard to this potential though. But i think, adding this feature now, and considering later exposing a way to control it would be beneficial, and prevents locking us into an opt-in behavior now.