@jasdel 2020-01-07
I've made a couple minor updates to the PR in order to be compatible with Go 1.8 through 1.5.