@dantidwell 2020-01-08
@jasdel Thanks for the review and the pre-go1.9 compatibility fixes!  I've pushed a couple of additional compatibility related changes to the tests, and now CI looks to be happy.  Regarding unbounded `sync.Map` growth, yeah I agree keeping tabs on that and potentially providing a way to manage it down the line is a good call.  On the bright side, the size of the cache is naturally bounded by the number of types you are persisting to the DB, which should be (hopefully) relatively small.